Refactor: Overlay Components
