+++ client/components/Overlay/Modals/Auth/RegisterModal.tsx
@@ -8,7 +8,7 @@ import { useSession } from 'next-auth/react';
 import { useRouter } from 'next/navigation';
 import React, { useEffect, useState } from 'react';
 import { toast } from 'react-hot-toast';
-import Modal from '../modal';
+import Modal from '../Modal';
 
 const RegisterModal = () => {
   const router = useRouter();
